breakup huge periodic task into two smaller utility functions
