+++ corehq/apps/analytics/utils.py
@@ -1,7 +1,16 @@
+import math
+import time
+import requests
+
 from django.conf import settings
 
+from corehq.util.metrics import metrics_gauge
 from corehq.apps.accounting.models import Subscription, BillingAccount
+from corehq.apps.es.users import UserES
 from corehq.apps.users.models import WebUser, CommCareUser
+
+
+MAX_API_RETRIES = 5
 
 
 def get_meta(request):
@@ -111,4 +120,174 @@ def get_instance_string():
     env = '' if instance == 'www' else instance + '_'
     return env
 
+
+def _delete_hubspot_contact(vid, retry_num=0):
+    """
+    Permanently deletes a Hubspot contact.
+    :param vid:  (the contact ID)
+    :param retry_num: the number of the current retry attempt
+    :return: boolean if contact was deleted
+    """
+    if retry_num > 0:
+        time.sleep(10)  # wait 10 seconds if this is another retry attempt
+
+    api_key = settings.ANALYTICS_IDS.get('HUBSPOT_API_KEY', None)
+    if api_key:
+        try:
+            req = requests.delete(
+                f'https://api.hubapi.com/contacts/v1/contact/vid/{vid}',
+                params={
+                    'hapikey': api_key,
+                }
+            )
+            if req.status_code == 404:
+                return False
+            req.raise_for_status()
+        except (ConnectionError, requests.exceptions.HTTPError):
+            metrics_gauge(
+                'commcare.hubspot_data.retry.delete_hubspot_contact',
+                1
+            )
+            if retry_num <= MAX_API_RETRIES:
+                return _delete_hubspot_contact(vid, retry_num + 1)
+        else:
+            return True
+    return False
+
+
+def _get_contact_ids_for_emails(list_of_emails, retry_num=0):
+    """
+    Gets a list of Contact IDs on Hubspot from a list of emails.
+    If an email in the list doesn't exist on Hubspot, it's simply ignored.
+    :param list_of_emails:
+    :param retry_num: the number of the current retry attempt
+    :return: list of contact ids
+    """
+    if retry_num > 0:
+        time.sleep(10)  # wait 10 seconds if this is another retry attempt
+
+    api_key = settings.ANALYTICS_IDS.get('HUBSPOT_API_KEY', None)
+    if api_key:
+        try:
+            req = requests.get(
+                "https://api.hubapi.com/contacts/v1/contact/emails/batch/",
+                params={
+                    'hapikey': api_key,
+                    'email': list_of_emails,
+                },
+            )
+            if req.status_code == 404:
+                return []
+            req.raise_for_status()
+        except (ConnectionError, requests.exceptions.HTTPError):
+            metrics_gauge(
+                'commcare.hubspot_data.retry.get_contact_ids_for_emails',
+                1
+            )
+            if retry_num <= MAX_API_RETRIES:
+                return _get_contact_ids_for_emails(list_of_emails, retry_num + 1)
+        else:
+            return req.json().keys()
+    return []
+
+
+def _get_contact_ids_for_email_domain(email_domain, retry_num=0):
+    """
+    Searches Hubspot for an email domain and returns the list of matching
+    contact IDs for that email domain.
+    :param email_domain:
+    :param retry_num: the number of the current retry attempt
+    :return: list of matching contact IDs
+    """
+    if retry_num > 0:
+        time.sleep(10)  # wait 10 seconds if this is another retry attempt
+
+    api_key = settings.ANALYTICS_IDS.get('HUBSPOT_API_KEY', None)
+    if api_key:
+        try:
+            req = requests.get(
+                "https://api.hubapi.com/contacts/v1/search/query",
+                params={
+                    'hapikey': api_key,
+                    'q': f'@{email_domain}',
+                },
+            )
+            if req.status_code == 404:
+                return []
+            req.raise_for_status()
+        except (ConnectionError, requests.exceptions.HTTPError):
+            metrics_gauge(
+                'commcare.hubspot_data.retry.get_contact_ids_for_email_domain',
+                1
+            )
+            if retry_num <= MAX_API_RETRIES:
+                return _get_contact_ids_for_email_domain(email_domain, retry_num + 1)
+        else:
+            return [contact.get('vid') for contact in req.json().get('contacts')]
+    return []
+
+
+def remove_blocked_email_domains_from_hubspot(stdout=None):
+    """
+    Removes contacts from Hubspot that emails matching our list of
+    blocked email domains.
+    :param stdout: the stdout of a management command (if applicable)
+    """
+    blocked_email_domains = get_blocked_hubspot_email_domains()
+    for email_domain in blocked_email_domains:
+        ids_to_delete = _get_contact_ids_for_email_domain(email_domain)
+        if stdout:
+            stdout.write(f"\n\nChecking EMAIL DOMAIN {email_domain}")
+            stdout.write(f"Found {len(ids_to_delete)} id(s) to delete.")
+        num_deleted = sum(_delete_hubspot_contact(vid) for vid in ids_to_delete)
+        metrics_gauge(
+            'commcare.hubspot_data.deleted_user.blocked_email_domain',
+            num_deleted,
+            tags={
+                'email_domain': email_domain,
+                'ids_deleted': ids_to_delete,
+            }
+        )
+
+
+def remove_blocked_domain_contacts_from_hubspot(stdout=None):
+    """
+    Removes contacts from Hubspot that are members of blocked domains.
+    :param stdout: the stdout of a management command (if applicable)
+    """
+    blocked_domains = get_blocked_hubspot_domains()
+    for domain in blocked_domains:
+        if stdout:
+            stdout.write(f"\n\nChecking DOMAIN {domain}...")
+        user_query = UserES().domain(domain).source(['email', 'username'])
+
+        total_users = user_query.count()
+        chunk_size = 30  # Hubspot recommends fewer than 100 emails per request
+        num_chunks = int(math.ceil(float(total_users) / float(chunk_size)))
+
+        for chunk in range(num_chunks):
+            blocked_users = (user_query
+                             .size(chunk_size)
+                             .start(chunk * chunk_size)
+                             .run()
+                             .hits)
+            blocked_emails = []
+            for user in blocked_users:
+                username = user.get('username')
+                user_email = user.get('email')
+                blocked_emails.append(username)
+                if user_email and user_email != username:
+                    blocked_emails.append(user_email)
+            ids_to_delete = _get_contact_ids_for_emails(set(blocked_emails))
+            if stdout:
+                stdout.write(f"Found {len(ids_to_delete)} id(s) to delete.")
+            num_deleted = sum(_delete_hubspot_contact(vid) for vid in ids_to_delete)
+            metrics_gauge(
+                'commcare.hubspot_data.deleted_user.blocked_domain',
+                num_deleted,
+                tags={
+                    'domain': domain,
+                    'ids_deleted': ids_to_delete,
+                }
+            )
